Networking day by day added
