+++ Coco/RegisterViewController.swift
@@ -9,7 +9,7 @@
 import UIKit
 import Firebase
 
-class RegisterViewController: UIViewController {
+class RegisterViewController: UIViewController, UITextFieldDelegate {
 
     @IBOutlet weak var userName: UITextField!
     @IBOutlet weak var password: UITextField!
@@ -18,7 +18,7 @@ class RegisterViewController: UIViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-
+        email.delegate = self
     }
         
     @IBAction func handleRegisterTapped(_ sender: UIButton) {
@@ -42,5 +42,10 @@ class RegisterViewController: UIViewController {
                 print("Registration error: \(String(describing: error))")
             }
         }
+    }
+    
+    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+        textField.resignFirstResponder()
+        return true
     }
 }
